sensor-generator/pyproject.toml
[project]
name = "sensor-generator"
version = "0.1.0"
description = "EKS ML Observability PoC — 합성 센서 데이터 생성기"
requires-python = ">=3.12"

dependencies = [
    "numpy>=2.0.0",
    "httpx>=0.27.0",
    "typer>=0.13.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
